% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JT-RCI.R
\name{JTRCI}
\alias{JTRCI}
\title{compute and plot Jacobson-Truax or reliable change indices}
\usage{
JTRCI(data = NA, pre = NA, post = NA, ppid = NA, group = NA,
  reliability = NA, higherIsBetter = F, indextype = "JT",
  JTcrit = "auto", normM = NA, normSD = NA, dysfM = NA,
  dysfSD = NA, plot = T, table = T, ...)
}
\arguments{
\item{data}{name of the dataframe containing pre and post data}

\item{pre}{name of the variable containing data at pre/baseline}

\item{post}{name of the variable containing data at post}

\item{ppid}{optional: name of variable containing participant identifying info. If not provided, the function will number participants/rows from 1: number of rows}

\item{group}{optional: name of variable containing grouping information (e.g. groups 1/2 or experimental/control)}

\item{reliability}{the value of the reliability estimate}

\item{higherIsBetter}{logical: if 'higherIsBetter = TRUE' the function assumes that a higher score indicates a better/more healthy/functional state}

\item{indextype}{choice of "JT" or "RCI". Indicates whether to obtain Jacobson-Truax clinicial significant change or reliable change indices respectively}

\item{JTcrit}{choice of "auto", "A", "B", "C". Indicates which criterion to use when obtaining Jacobson-Truax indices. 
When 'JTcrit = "auto"' the function will determine which criteria is likely the best fitting based on the available information (e.g. whether function/dysfunctional norm distributions are provided and whether these distributions overlap or not). When no norm-distributions are provided, 'JTcrit = "auto"' will result in the function using criterion A with the baseline distribution as a proxy for the dysfunctional distribution. 
Also note that when JTcrit is set to "B" or "C" but their required norm-data inputs are missing, it will output criterion A. When B or C is chosen but the other is a better fit, the function will create a message to inform the user about this.}

\item{normM}{mean value of the functional/healthy norm distribution}

\item{normSD}{SD value of the functional/healthy norm distribution}

\item{dysfM}{mean value of the dysfunctional norm distribution}

\item{dysfSD}{SD value of the dysfunctional norm distribution}

\item{plot}{logical, if plot = T a plot will be generated}

\item{table}{logical, if table = T a table will be generated}

\item{...}{further arguments to be passed by JTRCI() to the table or plot functions (i.e. you can enter all their parameters directly into JTRCI())}
}
\value{
JTRCIdf - a dataframe containing all values and results of the JT or RCI calculations
}
\description{
The JTRCI() function is the main function in the JTRCI package.
It requires data in wide format (one row per individual) and an estimate of the measure reliability to be given through parameter 'reliability ='.
}
\details{
The reliability estimate can be based on norm data. Alternatively, the internal reliability of the measure observed in the current data can be used. In that case, the user will first need to obtain their estimate of choice (alpha, omega, x-random splits splithalf reliability) using the raw (item level) data.

Obtaining reliable change indices is pretty straightforward: run the function with parameter 'indextype = "RCI"'. The JTRCI function's default setting is to obtain Jacobson-Truax indices ('indextype = "JT"'). The function is designed to provide guidance through the various choices a researcher has to make when applying the Jacobson-Truax method. It is recommended to pay attention to the various output messages.
}
\examples{
JTRCI(data = df, ppid = "ppid", pre = "pre", post = "post",  reliability = .8, indextype = "JT", JTcrit = "auto")
}
