% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRCI.R
\name{plotRCI}
\alias{plotRCI}
\title{display a plot of the obtained reliable change indices}
\usage{
plotRCI(data = JTRCIdf, addInfoLegend = c("yes", "no"),
  useGroups = F, facetplot = F, addJitter = F, xlab = "pre",
  ylab = "post", plottitle = "reliable change plot")
}
\arguments{
\item{data}{The data to use, default is JTRCIdf.}

\item{addInfoLegend}{choice of "yes" or "no": if yes a count of observations within each RCI class (or within each class and group) will be added to the plot legend}

\item{useGroups}{logical. If 'useGroups = T', dots are shaped and counts are given according seperately for each level of the group variable originally provided in the call to JTRCI()}

\item{facetplot}{logical. If 'facetplot = T', different groups are plotted in separate facets}

\item{addJitter}{logical. If 'addJitter = T', a small amount of random jitter is added to the plot}

\item{xlab}{label to use for the x-axis}

\item{ylab}{label to use for the y-axis}

\item{plottitle}{title to show at the top of the plot}
}
\value{
a plot showing reliable change indices
}
\description{
This function plotsdata stored in the JTRCIdf dataframe that is outputted by the JTRCI() function. When JTRCI() is run with it's default value 'plot = T', it will call upon either plotRCI() or plotJT() to create a plot. 
plotRCI() and plotJT() can also be called directly by the user to output plots after having obtained the JTRCIdf (by running JTRCI())
}
\examples{
plotRCI(xlab = "score pre", ylab = "score post", plottitle = "my reliable change plot", useGroups = F)
}
