% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableRCI.R
\name{tableRCI}
\alias{tableRCI}
\title{display a table summarizing reliable change indices}
\usage{
tableRCI(data = JTRCIdf, useGroups = NA)
}
\arguments{
\item{data}{The data to use, default is JTRCIdf.}

\item{useGroups}{if true, counts are given by each level of the group variable originally provided in the call to JTRCI()}
}
\value{
a table summarizing counts per classification label.
}
\description{
This function creates a table summarzing the number of observations for each label in the reliable change classification. It uses data stored in the JTRCIdf dataframe that is outputted by the JTRCI() function.
}
\examples{
tableRCI(useGroups = F)
#' @import data.table
}
